Fix Floorp split view drag and drop preventing tab dragging
